Fix unlink errors
unlink(/var/cache/demo.recras.nl/frontend/prod/template/.): Is a directory

+++ lib/cache/sfFileCache.class.php
@@ -149,7 +149,7 @@ class sfFileCache extends sfCache
     $result = true;
     foreach (new RecursiveIteratorIterator(new RecursiveDirectoryIterator($this->getOption('cache_dir'))) as $file)
     {
-      if (sfCache::ALL == $mode || !$this->isValid($file))
+      if ((sfCache::ALL == $mode || !$this->isValid($file)) && !is_dir($file))
       {
         $result = @unlink($file) && $result;
       }
